feat: Allow editing of shopping and todo list items
Introduces the ability to edit existing items in both the shopping list and todo list. This is achieved by introducing a new state variable `editingListItemId` and modifying the `handleAddList` function to handle updates when an `editingListItemId` is present. A new function `startEditingList` has also been added to populate the input fields with the item's text and set the `editingListItemId`.